Don't require cJSON in mosquitto_broker.h
