+++ plugins/dynamic-security/dynamic_security.h
@@ -22,6 +22,7 @@ Contributors:
 #include <uthash.h>
 #include "mosquitto.h"
 #include "mosquitto_broker.h"
+#include "mosquitto_control.h"
 #include "password_mosq.h"
 #include "base64_mosq.h"
