+++ src/broker_control.c
@@ -30,6 +30,7 @@ Contributors:
 #include "mosquitto.h"
 #include "mosquitto_broker_internal.h"
 #include "mosquitto_broker.h"
+#include "mosquitto_control.h"
 #include "mosquitto_plugin.h"
 #include "memory_mosq.h"
 #include "mqtt_protocol.h"
