+++ src/control_common.c
@@ -2,6 +2,7 @@
 
 #include <mqtt_protocol.h>
 #include <mosquitto_broker.h>
+#include <mosquitto_control.h>
 
 #include <stdlib.h>
 #include <string.h>
